pyproject.toml
[project]
name = "qr-gen-cli"
version = "0.1.0"
description = "CLI tool for generating branded QR codes"
authors = [{name = "Orlando Bruno"}]
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "qrcode[pil]>=7.4.2",
    "pillow>=10.0.0",
]

[project.scripts]
qrgen = "qr_gen_cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
